test: AppExceptionAdviceTest 수정
